Refactor project structure, add config and logging
