+++ .gitignore
@@ -5,5 +5,9 @@ __pycache__/
 # Virtual environment
 venv/
 
+# Logs and reports
+*.log
+*.report
+
 # macOS
 .DS_Store
