+++ seeding/models/__init__.py
@@ -1,0 +1,3 @@
+from .data_models import OriginalImage, ObjectImage, AllClassImage
+
+__all__ = ["OriginalImage", "ObjectImage", "AllClassImage"]
