+++ seeding/processing/__init__.py
@@ -1,0 +1,3 @@
+from .image_processor import ImageProcessor
+
+__all__ = ["ImageProcessor"]
