+++ seeding/ui/__init__.py
@@ -1,0 +1,4 @@
+from .main_window import ImageEditor
+from .tree_widget import LayerTreeWidget
+
+__all__ = ["ImageEditor", "LayerTreeWidget"]
